Style GUI with CSS files and an image
The application provides basic functions as a chatbot but has basic GUI.

Having a better GUI potentially increthe ases favour of users towards
the chatbot.

Let's create new css files to style the main window and dialog box with
colours and modify their fonts. Let's add an image to become the icon of
the application.